Rename menu -> megamenu
